+++ code/lib/cli-storybook/src/automigrate/fixes/addon-storysource-code-panel.ts
